+++ src/AutoUtils.c
@@ -1,33 +1,4 @@
 #include "AutoUtils.h"
-#include "encoder.h"
-
-
-PIDHandle* initPID(float kp, float ki, float kd, float target, float error_cap)
-{
-	PIDHandle* handle = NULL;
-	handle = malloc(sizeof(PIDHandle));
-	memset(handle, 0, sizeof(PIDHandle));
-	handle->kp = kp;
-	handle->ki = ki;
-	handle->ki = ki;
-	handle->error_cap = error_cap;
-	handle->target = target;
-	handle->integral = 0;
-	handle->last_error = 0.0;
-	return handle;
-}
-
-void freePID(PIDHandle* handle)
-{
-	free(handle);
-}
-
-void setPIDTarget(PIDHandle* handle, float target)
-{
-	handle->integral = 0;
-	handle->target = target;
-}
-
 
 void initAutoGlobals()
 {
@@ -46,35 +17,6 @@ void initAutoGlobals()
 	y_pid = initPID(0.5,0.0,0.0,0.0,10.0);
 	x_pid = initPID(0.5,0.0,0.0,0.0,10.0);
 	arm_macro_mutex = false;
-}
-
-float calculatePID(PIDHandle* handle, float position)
-{
-	if(handle == NULL) return 0.0;
-	float result = 0.0;
-	float pid_error = position - handle->target;
-	float derivative = 0.0;
-
-
-	//Integral
-	if( handle->ki != 0 )
-	{
-	// If we are inside controlable window then integrate the error
-	if( abs(pid_error) < handle->error_cap || handle->error_cap == 0.0)
-	  handle->integral = handle->integral + pid_error;
-	else
-	  handle->integral = 0;
-	}
-
-	// calculate the derivative
-	derivative = pid_error - handle->last_error;
-	handle->last_error  = pid_error;
-
-	//printf("%5.3f %5.3f %5.3f\n",pid_error, (handle->kp * pid_error), (handle->ki * handle->integral));
-
-	result = (handle->kp * pid_error) + (handle->ki * handle->integral) + (handle->kd * derivative);
-
-	return result;
 }
 
 void turnBase(short power)
@@ -233,17 +175,22 @@ void setAutoDriveMode(AutoDriveMode adm)
 	drive_mode = adm;
 }
 
-void pickUpCubeMacroThread()
+void pickUpUpCubeMacroThread()
 {
 	arm_macro_mutex = true;
-	delay(500);
-	setArmTarget(CLAW_GROUND_CUBE);
-
-	delay(1500);
 	clawPower(-127);
 	delay(750);
 	clawPower(-110);
 	setArmTarget(-999);
+	arm_macro_mutex = false;
+	return;
+}
+
+void pickUpDownCubeMacroThread()
+{
+	arm_macro_mutex = true;
+	delay(500);
+	setArmTarget(CLAW_GROUND_CUBE);
 	arm_macro_mutex = false;
 	return;
 }
@@ -309,12 +256,21 @@ void pickUpUpMacro()
 	}
 }
 
-void pickUpCubeMacro()
+void pickUpUpCubeMacro()
 {
 	//TODO Make less stupid and add safety (Cant run more than 1 and cancel mode)
 	if(arm_macro_mutex == false)
 	{
-		taskCreate(pickUpCubeMacroThread,  TASK_DEFAULT_STACK_SIZE, NULL, TASK_PRIORITY_DEFAULT);
+		taskCreate(pickUpUpCubeMacroThread,  TASK_DEFAULT_STACK_SIZE, NULL, TASK_PRIORITY_DEFAULT);
+	}
+}
+
+void pickUpDownCubeMacro()
+{
+	//TODO Make less stupid and add safety (Cant run more than 1 and cancel mode)
+	if(arm_macro_mutex == false)
+	{
+		taskCreate(pickUpDownCubeMacroThread,  TASK_DEFAULT_STACK_SIZE, NULL, TASK_PRIORITY_DEFAULT);
 	}
 }
